third_party/blink/web_tests/external/wpt/trusted-types/TrustedTypePolicyFactory-getAttributeType-svg-expected.txt

This is a testharness.js-based test.
[FAIL] trustedTypes.getAttributeType svg script[href]
  assert_equals: expected (string) "TrustedScriptURL" but got (object) null
[FAIL] trustedTypes.getAttributeType svg script[href] xlink href
  assert_equals: expected (string) "TrustedScriptURL" but got (object) null
Harness: the test ran to completion.
